add extensions for UI Tests, remove print statements
